requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
aiofiles==23.1.0
PyYAML>=6.0.1
pytest==7.2.0
pytest-asyncio==0.21.0
openai==1.64.0
httpx==0.23.3
faiss-cpu==1.10.0
numpy==2.2.3
python-dotenv==1.0.1
psutil==7.0.0